+++ sdk/resourcemanager/applicationinsights/armapplicationinsights/zz_generated_workbooks_client.go
@@ -57,9 +57,9 @@ func NewWorkbooksClient(subscriptionID string, credential azcore.TokenCredential
 
 // CreateOrUpdate - Create a new workbook.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-08-01
+// Generated from API version 2022-04-01
 // resourceGroupName - The name of the resource group. The name is case insensitive.
-// resourceName - The name of the Application Insights component resource.
+// resourceName - The name of the resource.
 // workbookProperties - Properties that need to be specified to create a new workbook.
 // options - WorkbooksClientCreateOrUpdateOptions contains the optional parameters for the WorkbooksClient.CreateOrUpdate
 // method.
@@ -101,7 +101,7 @@ func (client *WorkbooksClient) createOrUpdateCreateRequest(ctx context.Context, 
 	if options != nil && options.SourceID != nil {
 		reqQP.Set("sourceId", *options.SourceID)
 	}
-	reqQP.Set("api-version", "2021-08-01")
+	reqQP.Set("api-version", "2022-04-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, workbookProperties)
@@ -118,9 +118,9 @@ func (client *WorkbooksClient) createOrUpdateHandleResponse(resp *http.Response)
 
 // Delete - Delete a workbook.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-08-01
+// Generated from API version 2022-04-01
 // resourceGroupName - The name of the resource group. The name is case insensitive.
-// resourceName - The name of the Application Insights component resource.
+// resourceName - The name of the resource.
 // options - WorkbooksClientDeleteOptions contains the optional parameters for the WorkbooksClient.Delete method.
 func (client *WorkbooksClient) Delete(ctx context.Context, resourceGroupName string, resourceName string, options *WorkbooksClientDeleteOptions) (WorkbooksClientDeleteResponse, error) {
 	req, err := client.deleteCreateRequest(ctx, resourceGroupName, resourceName, options)
@@ -157,7 +157,7 @@ func (client *WorkbooksClient) deleteCreateRequest(ctx context.Context, resource
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-08-01")
+	reqQP.Set("api-version", "2022-04-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -165,9 +165,9 @@ func (client *WorkbooksClient) deleteCreateRequest(ctx context.Context, resource
 
 // Get - Get a single workbook by its resourceName.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-08-01
+// Generated from API version 2022-04-01
 // resourceGroupName - The name of the resource group. The name is case insensitive.
-// resourceName - The name of the Application Insights component resource.
+// resourceName - The name of the resource.
 // options - WorkbooksClientGetOptions contains the optional parameters for the WorkbooksClient.Get method.
 func (client *WorkbooksClient) Get(ctx context.Context, resourceGroupName string, resourceName string, options *WorkbooksClientGetOptions) (WorkbooksClientGetResponse, error) {
 	req, err := client.getCreateRequest(ctx, resourceGroupName, resourceName, options)
@@ -204,7 +204,7 @@ func (client *WorkbooksClient) getCreateRequest(ctx context.Context, resourceGro
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-08-01")
+	reqQP.Set("api-version", "2022-04-01")
 	if options != nil && options.CanFetchContent != nil {
 		reqQP.Set("canFetchContent", strconv.FormatBool(*options.CanFetchContent))
 	}
@@ -224,7 +224,7 @@ func (client *WorkbooksClient) getHandleResponse(resp *http.Response) (Workbooks
 
 // NewListByResourceGroupPager - Get all Workbooks defined within a specified resource group and category.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-08-01
+// Generated from API version 2022-04-01
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // category - Category of workbook to return.
 // options - WorkbooksClientListByResourceGroupOptions contains the optional parameters for the WorkbooksClient.ListByResourceGroup
@@ -283,7 +283,7 @@ func (client *WorkbooksClient) listByResourceGroupCreateRequest(ctx context.Cont
 	if options != nil && options.CanFetchContent != nil {
 		reqQP.Set("canFetchContent", strconv.FormatBool(*options.CanFetchContent))
 	}
-	reqQP.Set("api-version", "2021-08-01")
+	reqQP.Set("api-version", "2022-04-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -300,7 +300,7 @@ func (client *WorkbooksClient) listByResourceGroupHandleResponse(resp *http.Resp
 
 // NewListBySubscriptionPager - Get all Workbooks defined within a specified subscription and category.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-08-01
+// Generated from API version 2022-04-01
 // category - Category of workbook to return.
 // options - WorkbooksClientListBySubscriptionOptions contains the optional parameters for the WorkbooksClient.ListBySubscription
 // method.
@@ -351,7 +351,7 @@ func (client *WorkbooksClient) listBySubscriptionCreateRequest(ctx context.Conte
 	if options != nil && options.CanFetchContent != nil {
 		reqQP.Set("canFetchContent", strconv.FormatBool(*options.CanFetchContent))
 	}
-	reqQP.Set("api-version", "2021-08-01")
+	reqQP.Set("api-version", "2022-04-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -368,9 +368,9 @@ func (client *WorkbooksClient) listBySubscriptionHandleResponse(resp *http.Respo
 
 // RevisionGet - Get a single workbook revision defined by its revisionId.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-08-01
+// Generated from API version 2022-04-01
 // resourceGroupName - The name of the resource group. The name is case insensitive.
-// resourceName - The name of the Application Insights component resource.
+// resourceName - The name of the resource.
 // revisionID - The id of the workbook's revision.
 // options - WorkbooksClientRevisionGetOptions contains the optional parameters for the WorkbooksClient.RevisionGet method.
 func (client *WorkbooksClient) RevisionGet(ctx context.Context, resourceGroupName string, resourceName string, revisionID string, options *WorkbooksClientRevisionGetOptions) (WorkbooksClientRevisionGetResponse, error) {
@@ -412,7 +412,7 @@ func (client *WorkbooksClient) revisionGetCreateRequest(ctx context.Context, res
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-08-01")
+	reqQP.Set("api-version", "2022-04-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -429,9 +429,9 @@ func (client *WorkbooksClient) revisionGetHandleResponse(resp *http.Response) (W
 
 // NewRevisionsListPager - Get the revisions for the workbook defined by its resourceName.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-08-01
+// Generated from API version 2022-04-01
 // resourceGroupName - The name of the resource group. The name is case insensitive.
-// resourceName - The name of the Application Insights component resource.
+// resourceName - The name of the resource.
 // options - WorkbooksClientRevisionsListOptions contains the optional parameters for the WorkbooksClient.RevisionsList method.
 func (client *WorkbooksClient) NewRevisionsListPager(resourceGroupName string, resourceName string, options *WorkbooksClientRevisionsListOptions) *runtime.Pager[WorkbooksClientRevisionsListResponse] {
 	return runtime.NewPager(runtime.PagingHandler[WorkbooksClientRevisionsListResponse]{
@@ -481,7 +481,7 @@ func (client *WorkbooksClient) revisionsListCreateRequest(ctx context.Context, r
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-08-01")
+	reqQP.Set("api-version", "2022-04-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -498,9 +498,9 @@ func (client *WorkbooksClient) revisionsListHandleResponse(resp *http.Response) 
 
 // Update - Updates a workbook that has already been added.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-08-01
+// Generated from API version 2022-04-01
 // resourceGroupName - The name of the resource group. The name is case insensitive.
-// resourceName - The name of the Application Insights component resource.
+// resourceName - The name of the resource.
 // options - WorkbooksClientUpdateOptions contains the optional parameters for the WorkbooksClient.Update method.
 func (client *WorkbooksClient) Update(ctx context.Context, resourceGroupName string, resourceName string, options *WorkbooksClientUpdateOptions) (WorkbooksClientUpdateResponse, error) {
 	req, err := client.updateCreateRequest(ctx, resourceGroupName, resourceName, options)
@@ -540,7 +540,7 @@ func (client *WorkbooksClient) updateCreateRequest(ctx context.Context, resource
 	if options != nil && options.SourceID != nil {
 		reqQP.Set("sourceId", *options.SourceID)
 	}
-	reqQP.Set("api-version", "2021-08-01")
+	reqQP.Set("api-version", "2022-04-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if options != nil && options.WorkbookUpdateParameters != nil {